GPU: Fix fence timeout issue due to incorrect softstop handling [1/1]
PD#SWPL-143168

Problem:
A corner case race scenario was found where an atom of context B was
queued up for execution on a job slot, after a softstop was already
issued for an atom of context A running on that job slot.
When softstop command completed and JOB_IRQ was raised for the atom of
context A, atom of context B was removed from the HEAD_NEXT and returned
to the Job scheduler. However the removal was done without marking
the context B as blocked, which later caused a problem when the atom
returned to the Job scheduler. The context B was not added back to
the pullable list since it didn't transition from blocked to unblocked.
Consequently atom of context B didn't get submitted again to the GPU,
which eventually manifested as a fence timeout issue.

Solution:
To address the fence timeout issue, this commit ensures that the
required bookkeeping is done for the aforementioned scenario when
removing an atom from HEAD_NEXT after the sofstop completion.

Verify:
AndroidS valhall r44p0

Change-Id: Ib97d7e16a3917de1c72fe8a116d6793c689cd78b
Signed-off-by: dezhi kong <dezhi.kong@amlogic.com>